+++ release-notes.md
@@ -1,7 +1,28 @@
 # Version
-this verison = 2.16.1
-previous version 2.16.0
+this verison = 2.17.0
+previous version 2.16.1
 
+##  New Features
+1. [Create Objects with Factory Functions](https://github.com/vishalz/hello-js/issues/104)
+1. [basic.30.objects methods #97](https://github.com/vishalz/hello-js/issues/97)
+1. [Add closure concept to basic.30.scope #92](https://github.com/vishalz/hello-js/issues/92)
+1. [Creating Objects using literal notation #94](https://github.com/vishalz/hello-js/issues/94)
+1. [basic.30.objects add test for hasOwnProperty method #107](https://github.com/vishalz/hello-js/issues/107)
+1. [basic.30.objects use objects specifiers to pass parameters to functions #106](https://github.com/vishalz/hello-js/issues/106)
+1. [Objects have properties (simple types, methods and other Objects) #95](https://github.com/vishalz/hello-js/issues/95)
+
+
+## Fixed Issues
+
+1. [fix test for var str10 in basic.30.strings #88](https://github.com/vishalz/hello-js/ssues/88)
+1. [Fix v17 test in basic.30.strings #89](https://github.com/vishalz/hello-js/issues/89)
+1. [Cleanup Concepts.md #87](https://github.com/vishalz/hello-js/issues/87)
+1. [Concepts.md file incorrect #93](https://github.com/vishalz/hello-js/issues/93)
+1. [Fix CONCEPTS.md at basic.30.strings section #90](https://github.com/vishalz/hello-js/issues/90)
+
+
+
+#  2.16.1
 ##  New Features
 
 ## Fixed Issues